End of Section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -26,7 +27,7 @@ public class AuctionsController : ControllerBase
     }
 
     // api/auctions
-    [HttpGet(Name = "auctions.getAll")]
+    [HttpGet]
     public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string date)
     {
         // var auctions = await _context.Auctions.Include(x => x.Item)
@@ -47,7 +48,7 @@ public class AuctionsController : ControllerBase
     }
 
     // api/auctions/{id}
-    [HttpGet("{id}", Name = "auctions.getById")]
+    [HttpGet("{id}")]
     public async Task<ActionResult<AuctionDto>> GetAuctionById(Guid id)
     {
         var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x => x.Id == id);
@@ -58,12 +59,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
-    [HttpPost(Name = "auctions.create")]
+    [Authorize]
+    [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createAuctionDto);
-        // TODO: add current users as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
 
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -80,13 +82,16 @@ public class AuctionsController : ControllerBase
 
     }
 
-    [HttpPut("{id}", Name = "auctions.update")]
+    [Authorize]
+    [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x => x.Id == id);
         if (auction == null) return NotFound();
 
-        // TODO: check seller = username
+       
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -99,14 +104,15 @@ public class AuctionsController : ControllerBase
         if (result) return Ok();
         return BadRequest("Failed to update auction");
     }
-
-    [HttpDelete("{id}", Name = "auctions.delete")]
+    [Authorize]
+    [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
         if (auction == null) return NotFound();
 
-        // TODO: Check seller == username
+       
+        if(auction.Seller != User.Identity.Name) return Forbid();
         _context.Auctions.Remove(auction);
         await _publishEndpoint.Publish(new AuctionDeleted() { Id = auction.Id.ToString() });
         var result = await _context.SaveChangesAsync() > 0;
